refactor: remove modal module
