+++ apps/web/pages/_app.tsx
@@ -4,7 +4,7 @@ import { RecoilRoot } from 'recoil'
 import { MachineProvider } from '@codelab/frontend'
 import { appMachine } from '@codelab/modules/app-stories'
 import { AppLayoutContainer } from '@codelab/modules/layout-stories'
-import { ModalContainer } from '@codelab/modules/modal-stories'
+import { UserLoginModal, UserSignupModal } from '@codelab/modules/user-stories'
 
 require('highlight.js/styles/monokai-sublime.css')
 require('antd/dist/antd.css')
@@ -17,7 +17,8 @@ const App = ({ children }: PropsWithChildren<any>) => {
     <>
       {typeof window === 'undefined' ? null : (
         <>
-          <ModalContainer />
+          <UserSignupModal />
+          <UserLoginModal />
           <AppLayoutContainer>{children}</AppLayoutContainer>
         </>
       )}
