--- libs/modules/app-stories/src/view/index.ts
@@ -1,4 +1,0 @@
-export * from '../../../header-stories/src/HeaderMenu'
-export * from '../../../sidebar-stories/src/SidebarMenu'
-export * from '../../../modal-stories/src/Modal'
-export * from '../../../modal-stories/src/ModalContainer'
