+++ libs/modules/grid-stories/src/store/machine-grid.ts
@@ -64,7 +64,7 @@ export const createGridMachine = () => {
               } = event
 
               console.log(oldItem, newItem)
-              console.log('onDragStop')
+              console.log('onResizeStop')
             },
           },
         },
