--- libs/modules/modal-stories/README.md
@@ -1,7 +1,0 @@
-# modules-modal-stories
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test modules-modal-stories` to execute the unit tests via [Jest](https://jestjs.io).
