--- libs/modules/modal-stories/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  displayName: 'modules-modal-stories',
-  preset: '../../../jest.preset.js',
-  transform: {
-    '^.+\\.[tj]sx?$': [
-      'babel-jest',
-      { cwd: __dirname, configFile: './babel-jest.config.json' },
-    ],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/modules/modal-stories',
-}
