--- libs/modules/modal-stories/src/index.ts
@@ -1,2 +1,0 @@
-export * from './Modal'
-export * from './ModalContainer'
