+++ libs/modules/user-stories/src/useCases/userLogin/index.ts
@@ -1,3 +1,4 @@
 export * from './UserLoginButtonContainer'
 export * from './UserLoginForm'
 export * from './UserLoginState'
+export * from './UserLoginModal'
