+++ libs/modules/user-stories/src/useCases/userSignup/index.ts
@@ -2,3 +2,4 @@ export * from './UserSignupButtonContainer'
 export * from './UserSignupForm'
 export * from './UserSignupState'
 export * from './UserSignupServices'
+export * from './UserSignupModal'
