+++ nx.json
@@ -169,9 +169,6 @@
     "modules-layout-stories": {
       "tags": []
     },
-    "modules-modal-stories": {
-      "tags": []
-    },
     "modules-sidebar-stories": {
       "tags": []
     },
